requirements.txt
streamlit>=1.28.0
openai>=1.3.0
streamlit-audiorecorder>=0.1.0
audio-recorder-streamlit>=0.0.8
pydub>=0.25.1
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0 